Sync

+++ Slim/Utils/PluginDownloader.pm
@@ -1,5 +1,10 @@
 package Slim::Utils::PluginDownloader;
 
+# Logitech Media Server Copyright 2001-2020 Logitech.
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License,
+# version 2.
+
 # Plugins are downloaded to <cachedir>/DownloadedPlugins and then extracted to <cachedir>/InstalledPlugins/Plugins/
 #
 # Plugins zip files should not include any additional path information - i.e. they include the install.xml file at the top level
@@ -86,7 +91,7 @@ sub extract {
 		if (-d $targetDir) {
 
 			main::INFOLOG && $log->info("removing existing $targetDir");
-		
+
 			rmtree $targetDir;
 		}
 
@@ -112,16 +117,16 @@ sub extract {
 
 			# ignore additional directory information in zip
 			for my $search ("Plugins/$plugin/", "$plugin/") {
-				
+
 				if ( $zip->membersMatching("^$search") ) {
-					
+
 					$source = $search;
 					last;
 				}
 			}
-			
+
 			if ( ($zipstatus = $zip->extractTree($source, "$targetDir/")) == Archive::Zip::AZ_OK() ) {
-					
+
 				main::INFOLOG && $log->info("extracted $plugin to $targetDir");
 
 			} else {
@@ -151,15 +156,28 @@ sub install {
 
 	my $file = catdir($downloadTo, "$name.zip");
 
-	unlink $file;
+	if (-r $file) {
+		my $digest = _getDigest($file);
 
-	my $http = Slim::Networking::SimpleAsyncHTTP->new( \&_downloadDone, \&_downloadError, { saveAs => $file, args => $args } );
-	
-	main::INFOLOG && $log->info("install - downloading $name from $url");
+		if ($args->{sha} ne $digest) {
+			main::INFOLOG && $log->is_info && $log->info(sprintf("found existing download, but digest does not match %s - %s will be re-downloaded: expected %s, got %s", $file, $name, $args->{sha}, $digest));
+			unlink $file;
+		}
+	}
 
-	$downloading++;
+	if (-r $file) {
+		main::DEBUGLOG && $log->is_debug && $log->debug("found existing download, digest matches - don't re-download: $name");
+		_installDownload($file, $args);
+	}
+	else {
+		my $http = Slim::Networking::SimpleAsyncHTTP->new( \&_downloadDone, \&_downloadError, { saveAs => $file, args => $args } );
 
-	$http->get($url);
+		main::INFOLOG && $log->info("install - downloading $name from $url");
+
+		$downloading++;
+
+		$http->get($url);
+	}
 }
 
 sub _downloadDone {
@@ -169,31 +187,41 @@ sub _downloadDone {
 	my $args  = $http->params('args');
 
 	my $name  = $args->{'name'};
-	my $digest= $args->{'sha'};
-	my $url   = $http->url;
 
 	main::INFOLOG && $log->info("downloaded $name to $file");
 
 	$downloading--;
 
+	_installDownload($file, $args);
+}
+
+sub _installDownload {
+	my ($file, $args) = @_;
+
+	my $name  = $args->{'name'};
+	my $digest= $args->{'sha'};
+
 	if (-r $file) {
 
-		my $sha1 = Digest::SHA1->new;
-		
-		open my $fh, '<', $file;
+		my $gotDigest = _getDigest($file);
 
-		binmode $fh;
-		
-		$sha1->addfile($fh);
-		
-		close $fh;
-		
-		if ($digest ne $sha1->hexdigest) {
-			
-			$log->warn("digest does not match $file - $name will not be installed");
+		if ($digest ne $gotDigest) {
 
-			unlink $file;
-			
+			$log->error("digest does not match $file - $name will not be installed: expected $digest, got " . $gotDigest);
+
+			if (main::DEBUGLOG && $log->is_debug && (my $cacheDir = preferences('server')->get('cachedir'))) {
+				require File::Basename;
+				require File::Copy;
+
+				my ($filename, undef, $ext) = File::Basename::fileparse($file, '.zip');
+				my $backup = catfile($cacheDir, $filename . '-' . $gotDigest . $ext);
+				$log->debug("Moving downloaded file for you to investigate:\n  from: $file\n  to:   $backup");
+				File::Copy::move($file, $backup);
+			}
+			else {
+				unlink $file;
+			}
+
 		} else {
 
 			main::INFOLOG && $log->info("digest matches - scheduling $name for install on restart");
@@ -201,6 +229,22 @@ sub _downloadDone {
 			$prefs->set($name, 'needs-install');
 		}
 	}
+}
+
+sub _getDigest {
+	my ($file) = @_;
+
+	my $sha1 = Digest::SHA1->new;
+
+	open my $fh, '<', $file or $log->error("Failed to open $file");
+
+	binmode $fh;
+
+	$sha1->addfile($fh);
+
+	close $fh;
+
+	return $sha1->hexdigest;
 }
 
 sub _downloadError {
@@ -280,7 +324,7 @@ sub _handleResponse {
 		Slim::Utils::PluginManager->message(
 			sprintf( "%s (%s)", Slim::Utils::Strings::string('PLUGINS_UPDATES_AVAILABLE'), $updates )
 		);
-		
+
 		# $updates is only set if we don't want to auto-update
 		return;
 	}
@@ -292,13 +336,13 @@ sub _handleResponse {
 			main::INFOLOG && $log->info("ignoring response - $plugin already pending action: " . $prefs->get($plugin));
 			next;
 		}
-		
+
 		my $entry = $actions->{$plugin};
 
 		if ($entry->{'action'} eq 'install' && $entry->{'url'} && $entry->{'sha'}) {
 
 			$class->install({ name => $plugin, url => $entry->{'url'}, sha => $entry->{'sha'} });
-							 
+
 		} elsif ($entry->{'action'} eq 'uninstall') {
 
 			$class->uninstall($plugin);
